+++ packages/web-components/src/components/dotcom-shell/__stories__/dotcom-shell.stories.ts
@@ -1153,6 +1153,112 @@ WithUniversalBanner.story = {
   },
 };
 
+export const WithoutShell = ({ parameters }) => {
+  const { masthead, universalBanner, tocLayout } = parameters?.props?.DotcomShell ?? {};
+
+  return html`
+    <style>
+      ${mastheadStyles}
+    </style>
+    ${universalBanner
+      ? html`
+          <dds-universal-banner image-width="4-col">
+            <dds-universal-banner-image slot="image" default-src="${images['4-col']}"></dds-universal-banner-image>
+            <dds-universal-banner-heading slot="heading">heading</dds-universal-banner-heading>
+            <dds-universal-banner-copy slot="copy">copy</dds-universal-banner-copy>
+            <dds-button-cta slot="cta" cta-type="local" kind="tertiary" href="https://www.example.com">
+              cta copy
+            </dds-button-cta>
+          </dds-universal-banner>
+        `
+      : ''}
+    ${masthead === 'L0'
+      ? html`
+          <dds-masthead-container id="masthead-container"></dds-masthead-container>
+        `
+      : html`
+          <dds-masthead-container id="masthead-container" .l1Data="${l1Data}"></dds-masthead-container>
+        `}
+    <main class="bx--content dds-ce-demo--ui-shell-content">
+      <div class="bx--grid">
+        <div class="bx--row">
+          <div class="bx--col-sm-8 bx--col-lg-16">
+            <div style="height: 400px; max-height: 75vh; padding: 1rem; background: #bbb;">
+              <h2>Leadspace</h2>
+            </div>
+            <dds-table-of-contents toc-layout="${tocLayout ?? ''}">
+              <div class="bx--tableofcontents__contents">
+                <a name="1" data-title="Section - 1"></a>
+                <h3 style="margin:4rem 0;">Section - 1</h3>
+                <div style="height: 400px; max-height: 75vh; padding: 1rem; background: #bbb;">
+                  <p>Lorem ipsum dolor sit amet</p>
+                </div>
+
+                <a name="2" data-title="Section - 2"></a>
+                <h3 style="margin:4rem 0;">Section - 2</h3>
+                <div style="height: 400px; max-height: 75vh; padding: 1rem; background: #bbb;">
+                  <p>Lorem ipsum dolor sit amet</p>
+                </div>
+
+                <a name="3" data-title="Section - 3"></a>
+                <h3 style="margin:4rem 0;">Section - 3</h3>
+                <div style="height: 400px; max-height: 75vh; padding: 1rem; background: #bbb;">
+                  <p>Lorem ipsum dolor sit amet</p>
+                </div>
+
+                <a name="4" data-title="Section - 4"></a>
+                <h3 style="margin:4rem 0;">Section - 4</h3>
+                <div style="height: 400px; max-height: 75vh; padding: 1rem; background: #bbb;">
+                  <p>Lorem ipsum dolor sit amet</p>
+                </div>
+
+                <a name="5" data-title="Section - 5"></a>
+                <h3 style="margin:4rem 0;">Section - 5</h3>
+                <div style="height: 400px; max-height: 75vh; padding: 1rem; background: #bbb;">
+                  <p>Lorem ipsum dolor sit amet</p>
+                </div>
+
+                <a name="6" data-title="Section - 6"></a>
+                <h3 style="margin:4rem 0;">Section - 6</h3>
+                <div style="height: 400px; max-height: 75vh; padding: 1rem; background: #bbb;">
+                  <p>Lorem ipsum dolor sit amet</p>
+                </div>
+
+                <a name="7" data-title="Section - 7"></a>
+                <h3 style="margin:4rem 0;">Section - 7</h3>
+                <div style="height: 400px; max-height: 75vh; padding: 1rem; background: #bbb;">
+                  <p>Lorem ipsum dolor sit amet</p>
+                </div>
+              </div>
+            </dds-table-of-contents>
+          </div>
+        </div>
+      </div>
+    </main>
+  `;
+};
+
+WithoutShell.story = {
+  name: 'Without Shell (Fallback Utility)',
+  parameters: {
+    knobs: {
+      DotcomShell: ({ groupId }) => ({
+        masthead: select('Masthead Version', ['L0', 'L1'], 'L0', groupId),
+        universalBanner: boolean('Has Universal Banner', false, groupId),
+        tocLayout: select('Table of Contents Layout', { Vertical: null, Horizontal: 'horizontal' }, null, groupId),
+      }),
+    },
+    propsSet: {
+      default: {
+        FooterComposite: {
+          disableLocaleButton: false,
+          langList: mockLangList,
+        },
+      },
+    },
+  },
+};
+
 export default {
   title: 'Components/Dotcom shell',
   decorators: [