+++ packages/storybook-addon-theme/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.23.0](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/main/packages/storybook-addon-theme/compare/@carbon/storybook-addon-theme@1.23.0-rc.1...@carbon/storybook-addon-theme@1.23.0) (2022-03-28)
+
+**Note:** Version bump only for package @carbon/storybook-addon-theme
+
+
+
+
+
+# [1.23.0-rc.1](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/main/packages/storybook-addon-theme/compare/@carbon/storybook-addon-theme@1.23.0-rc.0...@carbon/storybook-addon-theme@1.23.0-rc.1) (2022-03-17)
+
+**Note:** Version bump only for package @carbon/storybook-addon-theme
+
+
+
+
+
+# [1.23.0-rc.0](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/main/packages/storybook-addon-theme/compare/@carbon/storybook-addon-theme@1.22.0...@carbon/storybook-addon-theme@1.23.0-rc.0) (2022-03-16)
+
+**Note:** Version bump only for package @carbon/storybook-addon-theme
+
+
+
+
+
+# [1.22.0](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/main/packages/storybook-addon-theme/compare/@carbon/storybook-addon-theme@1.22.0-rc.3...@carbon/storybook-addon-theme@1.22.0) (2022-02-28)
+
+**Note:** Version bump only for package @carbon/storybook-addon-theme
+
+
+
+
+
 # [1.22.0-rc.3](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/main/packages/storybook-addon-theme/compare/@carbon/storybook-addon-theme@1.22.0-rc.2...@carbon/storybook-addon-theme@1.22.0-rc.3) (2022-02-25)
 
 **Note:** Version bump only for package @carbon/storybook-addon-theme